--- tests/mock_data/apply_relay_log_data.py
@@ -1,4 +1,0 @@
-from hexbytes import HexBytes
-
-apply_relay_event_logs = [{'logIndex': 1, 'transactionIndex': 0, 'transactionHash': HexBytes('0xb17c9a5e610b2ee3ddf2c5bfb6e26f26d9d75a11cde94f99b0be4f51999de59c'), 'blockHash': HexBytes('0xa5d76df4802af1dffde6317fc2c8e3a6e3a367926d7a2a75392026f2197c04ef'), 'blockNumber': 6, 'address': '0x9A8f367453B13cD30E9987F353Ad4b358A854895', 'data': '0x000000000000000000000000000000000000000000000000000000000000000a85af52eb9d1420e2fb96e970da8945f166952927f16656427557b5cf0d890a09', 'topics': [HexBytes('0xb114698a397e5f7a7c78e5e58f6f59d2c329c988c585a930db7bb099b3d016d6'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd')], 'type': 'mined'},
-                          {'logIndex': 1, 'transactionIndex': 0, 'transactionHash': HexBytes('0xafe0cad864a70d7bfb6d3310de0789094bd4cfe71931be49833f7959a102e311'), 'blockHash': HexBytes('0xe532eaebd98fbeb3548c7d122cf3fcf7fbb56b65ccb1f1ca78dd2cb6e18296bc'), 'blockNumber': 7, 'address': '0x9A8f367453B13cD30E9987F353Ad4b358A854895', 'data': '0x0000000000000000000000000000000000000000000000000de0b6b3a7640000aea53798eee98e90b0e3183638fa1687625bf606d2174c80fc4b416eff715435', 'topics': [HexBytes('0xb114698a397e5f7a7c78e5e58f6f59d2c329c988c585a930db7bb099b3d016d6'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd')], 'type': 'mined'}]
