--- tests/mock_data/relay_log_data.py
@@ -1,4 +1,0 @@
-from hexbytes import HexBytes
-
-relay_event_logs = [{'address': '0xE9118Ee0a63ABd22D3d3cDCeFF815f4343553823', 'blockHash': HexBytes('0xa8cc255a5b1dce315fe6fcddbe34a4812ea28d966bec010151fcd748dda2d6b2'), 'blockNumber': 5337358, 'data': '0x000000000000000000000000000000000000000000000000000000000000000a0000000000000000000000000000000000000000000000000000000000000064000000000000000000000000000000000000000000000000000000005ca5a633', 'logIndex': 2, 'removed': False, 'topics': [HexBytes('0xb77c820b3a0ee4da03c984a58bfe43cb27cd3297d424e1025014ce0b7de08cc4'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd')], 'transactionHash': HexBytes('0x85af52eb9d1420e2fb96e970da8945f166952927f16656427557b5cf0d890a09'), 'transactionIndex': 2}, {'address': '0xE9118Ee0a63ABd22D3d3cDCeFF815f4343553823', 'blockHash': HexBytes('0x6f4f96fe264715819f5fe9b19c6dadd13e5cfbdf9372e941b526beb9087efb0e'), 'blockNumber': 5337881, 'data': '0x0000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005ca5c0d9', 'logIndex': 13, 'removed': False, 'topics': [
-    HexBytes('0xb77c820b3a0ee4da03c984a58bfe43cb27cd3297d424e1025014ce0b7de08cc4'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd'), HexBytes('0x000000000000000000000000cda0d5e3f740f62767c0a91a10f0126e7362f4dd')], 'transactionHash': HexBytes('0xaea53798eee98e90b0e3183638fa1687625bf606d2174c80fc4b416eff715435'), 'transactionIndex': 48}, {'address': '0xE9118Ee0a63ABd22D3d3cDCeFF815f4343553823', 'blockHash': HexBytes('0xe3f9b6ef9eecb6efd03703278f0d1b15ce37cf4d1d7115499da18fa75df25f8a'), 'blockNumber': 5362311, 'data': '0x0000000000000000000000000000000000000000000000000de0b6b3a76400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005caac417', 'logIndex': 5, 'removed': False, 'topics': [HexBytes('0xb77c820b3a0ee4da03c984a58bfe43cb27cd3297d424e1025014ce0b7de08cc4'), HexBytes('0x000000000000000000000000d0d4e87c034d94e67feefaa445277d96766065bb'), HexBytes('0x000000000000000000000000016cb9eb8bda4cf3e333e5f79ed303d0a2725087')], 'transactionHash': HexBytes('0x0bf618cf7ccc752e7492f158d0b6b9399791a5d5a35b39bfc655b2c0a8bcb515'), 'transactionIndex': 20}]
